requirements.txt
python-dotenv
PyMuPDF
requests
beautifulsoup4
pillow
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
edge-tts==7.0.2
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.6.0
gTTS==2.5.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
propcache==0.3.1
PyMuPDF==1.25.5
requests==2.32.3
srt==3.5.3
tabulate==0.9.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
azure-cognitiveservices-vision-computervision
SQLAlchemy==2.0.36
PyMySQL==1.1.0
PyMySQL==1.1.0
cryptography==42.0.5
